requirements.txt
line-bot-sdk==2.4.2
Flask==2.3.3
Flask-SQLAlchemy==3.0.3
numpy==1.23.5
keras==2.12.0
Pillow==9.5.0
tensorflow-cpu==2.12.0
discord==2.3.2
discord.py==2.3.2
psycopg2==2.9.9
psycopg2-binary==2.9.5
emoji==2.10.1
python-dateutil==2.8.2
Pillow==9.5.0
